# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Veterinaria API",
        "description": "Documentación OpenAPI de la API de veterinaria.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/stats/dashboard": {
            "get": {
                "tags": [
                    "Estadísticas"
                ],
                "summary": "Dashboard principal",
                "description": "GET /stats/dashboard\nNúmeros rápidos para el panel principal",
                "operationId": "1976ae6d9660a18f788e5c2adaaeacad",
                "responses": {
                    "200": {
                        "description": "Resumen del dashboard",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardStats"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/vaccine-alerts": {
            "get": {
                "tags": [
                    "Estadísticas"
                ],
                "summary": "Alertas de vacunas",
                "description": "GET /stats/vaccine-alerts\nMascotas con vacuna vencida o próxima a vencer",
                "operationId": "198f8cd6affd66ba521687a27ede31d3",
                "responses": {
                    "200": {
                        "description": "Alertas de vacunas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/VaccineAlert"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/unvaccinated": {
            "get": {
                "tags": [
                    "Estadísticas"
                ],
                "summary": "Mascotas sin vacunas",
                "description": "GET /stats/unvaccinated\nMascotas sin ninguna vacuna registrada",
                "operationId": "608c5a7362f6963d790a27cc84737580",
                "responses": {
                    "200": {
                        "description": "Mascotas sin vacunas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UnvaccinatedPet"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/monthly-activity": {
            "get": {
                "tags": [
                    "Estadísticas"
                ],
                "summary": "Actividad mensual",
                "description": "GET /stats/monthly-activity\nMascotas nuevas y vacunas aplicadas por mes (últimos 12 meses)",
                "operationId": "39a4dda56b7290edafdb4dbf06518314",
                "responses": {
                    "200": {
                        "description": "Actividad mensual",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MonthlyActivity"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/species-distribution": {
            "get": {
                "tags": [
                    "Estadísticas"
                ],
                "summary": "Distribución por especie",
                "description": "GET /stats/species-distribution\nDistribución por especie",
                "operationId": "4aa4c862b277da80ca6b132df3634f66",
                "responses": {
                    "200": {
                        "description": "Distribución por especie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SpeciesDistribution"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Autentica al usuario y retorna un JWT válido.",
                "operationId": "e0c548fa9b5f5327b9e18821300bc7e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Invalida el token JWT actual.",
                "operationId": "b7f3adcb1bffd3a21330d7e7361c0e5f",
                "responses": {
                    "200": {
                        "description": "Cierre de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cierre de sesión exitoso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Usuario autenticado",
                "description": "Retorna los datos del usuario autenticado actual.",
                "operationId": "31037e6b0e8c45e8a7b55a2d16cef825",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthLoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@veterinaria.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    }
                },
                "type": "object"
            },
            "UserSummary": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Brayan Cárdenas"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@veterinaria.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "ADMIN"
                    }
                },
                "type": "object"
            },
            "AuthSuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Inicio de sesión exitoso"
                    },
                    "token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    },
                    "token_type": {
                        "type": "string",
                        "example": "bearer"
                    },
                    "expires_in": {
                        "type": "integer",
                        "example": 3600
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSummary"
                    }
                },
                "type": "object"
            },
            "PetSummary": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Firulais"
                    }
                },
                "type": "object"
            },
            "RoleList": {
                "type": "array",
                "items": {
                    "type": "string"
                },
                "example": [
                    "ADMIN"
                ]
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "dni": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "name": {
                        "type": "string",
                        "example": "Brayan Cárdenas"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@veterinaria.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "999888777",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Av. Central 123",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -12.0464,
                        "nullable": true
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -77.0428,
                        "nullable": true
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "roles": {
                        "$ref": "#/components/schemas/RoleList"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OwnerSummary": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "phone": {
                        "type": "string",
                        "example": "999888777",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "DashboardStats": {
                "properties": {
                    "total_pets": {
                        "type": "integer",
                        "example": 120
                    },
                    "total_owners": {
                        "type": "integer",
                        "example": 45
                    },
                    "vaccinated_this_month": {
                        "type": "integer",
                        "example": 17
                    },
                    "overdue_vaccines": {
                        "type": "integer",
                        "example": 8
                    },
                    "upcoming_in_30_days": {
                        "type": "integer",
                        "example": 12
                    },
                    "unvaccinated_count": {
                        "type": "integer",
                        "example": 6
                    }
                },
                "type": "object"
            },
            "MonthlyActivity": {
                "properties": {
                    "month": {
                        "type": "string",
                        "example": "May"
                    },
                    "month_key": {
                        "type": "string",
                        "example": "2026-05"
                    },
                    "new_pets": {
                        "type": "integer",
                        "example": 9
                    },
                    "vaccines_applied": {
                        "type": "integer",
                        "example": 14
                    }
                },
                "type": "object"
            },
            "SpeciesDistribution": {
                "properties": {
                    "species": {
                        "type": "string",
                        "example": "Canino"
                    },
                    "total": {
                        "type": "integer",
                        "example": 75
                    },
                    "percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 62.5
                    }
                },
                "type": "object"
            },
            "UnvaccinatedPet": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Firulais"
                    },
                    "species": {
                        "type": "string",
                        "example": "Canino"
                    },
                    "race": {
                        "type": "string",
                        "example": "Mestizo"
                    },
                    "status": {
                        "type": "string",
                        "example": "ADOPTADO"
                    },
                    "registered_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-18"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/OwnerSummary"
                    }
                },
                "type": "object"
            },
            "VaccineAlert": {
                "properties": {
                    "vaccine_id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "type": {
                        "type": "string",
                        "example": "Rabia"
                    },
                    "alert_type": {
                        "type": "string",
                        "example": "VENCIDA"
                    },
                    "days_diff": {
                        "type": "integer",
                        "example": -3
                    },
                    "application_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-18"
                    },
                    "next_vaccine_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-18"
                    },
                    "months_validity": {
                        "type": "integer",
                        "example": 12
                    },
                    "pet": {
                        "$ref": "#/components/schemas/PetSummary"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/OwnerSummary"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "No autenticado."
                    },
                    "status": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Los datos enviados no son válidos."
                    },
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "El campo email es obligatorio."
                                ]
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "El campo password es obligatorio."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación JWT usando Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}